requirements.txt
opencv-contrib-python~=4.5.5
pytesseract~=0.3.9
imutils~=0.5.4
Image~=1.5.33
numpy~=1.22.4
Flask~=2.1.2
werkzeug~=2.1.2
Pillow~=9.1.1
goslate~=1.5.2